collect and store recruit info
Collects and stores recruit information, allows the user to validate their inputs, calls on methods throughout. 
Input validation is performed via while loops that check to ensure conditions are met, and only conditions that need to be repeated are repeated (via staggered if/elif/else statements) and try/except blocks. 

Future development: 
Allowing coaches to input the values they're looking for.
Raising exceptions using coachs' values as parameters.
Uploading information to a pdf file for the coach to see. 
Formatting the pdf file. 
Integrating into a .HTML for better user presentation. 

Potential future applications:
Persistent files for users. 
Machine learning to identify the likelihood or recruit success or adjustment of the coach's parameters. 
Determining the top 10 percent of recruits that tried out. 
Recursive function to only allow a certain amount of recruits to register (code variant).